% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesRTMB_package.R
\docType{package}
\name{BayesRTMB-package}
\alias{BayesRTMB}
\alias{BayesRTMB-package}
\title{BayesRTMB: Bayesian Inference using RTMB}
\description{
A package for MCMC and MAP estimation using RTMB.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/norimune/BayesRTMB}
  \item \url{https://norimune.github.io/BayesRTMB/}
  \item Report bugs at \url{https://github.com/norimune/BayesRTMB/issues}
}

}
\keyword{internal}
