% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{bernoulli_logit_lpmf}
\alias{bernoulli_logit_lpmf}
\title{Bernoulli log-probability mass function with logit parameterization}
\usage{
bernoulli_logit_lpmf(x, mu)
}
\arguments{
\item{x}{Vector of binary outcomes (0 or 1).}

\item{mu}{Linear predictor (logit probability).}
}
\value{
The sum of the log-probability.
}
\description{
Bernoulli log-probability mass function with logit parameterization
}
