% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{bernoulli_lpmf}
\alias{bernoulli_lpmf}
\title{Bernoulli log-probability mass function}
\usage{
bernoulli_lpmf(x, prob)
}
\arguments{
\item{x}{Vector of binary outcomes (0 or 1).}

\item{prob}{Probability of success.}
}
\value{
The sum of the log-probability.
}
\description{
Bernoulli log-probability mass function
}
