% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{beta_lpdf}
\alias{beta_lpdf}
\title{Beta log-probability density function}
\usage{
beta_lpdf(x, a, b)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{a}{Shape parameter alpha.}

\item{b}{Shape parameter beta.}
}
\value{
The sum of the log-density.
}
\description{
Beta log-probability density function
}
