% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BigFive}
\alias{BigFive}
\title{Big Five Personality Traits Data}
\format{
A data frame with 10 rows and 20 columns:
\describe{
  \item{BF1}{Extraversion item 1}
  \item{BF2}{Neuroticism item 1}
  \item{BF3}{Openness item 1}
  \item{BF4}{Agreeableness item 1}
  \item{BF5}{Conscientiousness item 1}
  \item{BF6}{Extraversion item 2}
  \item{BF7}{Neuroticism item 2}
  \item{BF8}{Openness item 2}
  \item{BF9}{Agreeableness item 2}
  \item{BF10}{Conscientiousness item 2}
  \item{BF11}{Extraversion item 3}
  \item{BF12}{Neuroticism item 3}
  \item{BF13}{Openness item 3}
  \item{BF14}{Agreeableness item 3}
  \item{BF15}{Conscientiousness item 3}
  \item{BF16}{Extraversion item 4}
  \item{BF17}{Neuroticism item 4}
  \item{BF18}{Openness item 4}
  \item{BF19}{Agreeableness item 4}
  \item{BF20}{Conscientiousness item 4}
}
}
\source{
Originally collected by the package author.
}
\usage{
BigFive
}
\description{
A dataset containing responses to 20 items measuring the Big Five personality traits.
This dataset was originally collected by the package author.
The items are arranged in a cyclical order representing the five factors:
Extraversion, Neuroticism, Openness, Agreeableness, and Conscientiousness.
}
\keyword{datasets}
