% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{binomial_lpmf}
\alias{binomial_lpmf}
\title{Binomial log-probability mass function}
\usage{
binomial_lpmf(x, size, prob)
}
\arguments{
\item{x}{Vector of quantiles (number of successes).}

\item{size}{Number of trials.}

\item{prob}{Probability of success.}
}
\value{
The sum of the log-probability.
}
\description{
Binomial log-probability mass function
}
\keyword{internal}
