% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{bw_categorical_logit_lpmf}
\alias{bw_categorical_logit_lpmf}
\title{Best-Worst Categorical Logit Log-PMF (RTMB optimized)}
\usage{
bw_categorical_logit_lpmf(x, U, lambda = 1)
}
\arguments{
\item{x}{Vector of indices of the selected pairs.}

\item{U}{Vector of utilities for each item.}

\item{lambda}{Scaling parameter.}
}
\value{
Log-probability (scalar advector).
}
\description{
Best-Worst Categorical Logit Log-PMF (RTMB optimized)
}
