% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{bws_lpmf}
\alias{bws_lpmf}
\title{Best-Worst Scaling log-probability mass function}
\usage{
bws_lpmf(x, U, lambda = 1)
}
\arguments{
\item{x}{Integer. The index of the chosen best-worst pair (1 to C*(C-1)).}

\item{U}{A numeric vector of utilities for the items in the current choice set.}

\item{lambda}{A numeric scalar for the scale parameter (default is 1).}
}
\value{
The log-probability of the observed best-worst choice.
}
\description{
Best-Worst Scaling log-probability mass function
}
