% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{categorical_logit_lpmf}
\alias{categorical_logit_lpmf}
\title{Categorical log-probability mass function with logit parameterization}
\usage{
categorical_logit_lpmf(x, eta)
}
\arguments{
\item{x}{Integer or integer vector. The observed category index.}

\item{eta}{A numeric vector of linear predictors (unnormalized log-probabilities).}
}
\value{
The sum of the log-probability.
}
\description{
Categorical log-probability mass function with logit parameterization
}
\keyword{internal}
