% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{categorical_lpmf}
\alias{categorical_lpmf}
\title{Categorical log-probability mass function}
\usage{
categorical_lpmf(x, prob)
}
\arguments{
\item{x}{Vector of categorical outcomes.}

\item{prob}{Vector or matrix of probabilities.}
}
\value{
The sum of the log-probability.
}
\description{
Categorical log-probability mass function
}
\keyword{internal}
