% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{cauchy_lpdf}
\alias{cauchy_lpdf}
\title{Cauchy log-probability density function}
\usage{
cauchy_lpdf(x, location, scale)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{location}{Location parameter.}

\item{scale}{Scale parameter.}
}
\value{
The sum of the log-density.
}
\description{
Cauchy log-probability density function
}
\keyword{internal}
