% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{centered_multi_normal_lpdf}
\alias{centered_multi_normal_lpdf}
\title{Centered / Centered matrix multivariate normal log-probability density function}
\usage{
centered_multi_normal_lpdf(x, sigma, K = NULL)
}
\arguments{
\item{x}{Vector or matrix of quantiles. If a matrix, it assumes each column sums to zero.}

\item{sigma}{Standard deviation parameter. Can be a scalar, or a vector of length ncol(x) if x is a matrix.}

\item{K}{Dimension. If NULL, inferred from the length or nrow of x.}
}
\value{
The log-density.
}
\description{
Centered / Centered matrix multivariate normal log-probability density function
}
\keyword{internal}
