% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{centered_tri_multi_normal_lpdf}
\alias{centered_tri_multi_normal_lpdf}
\title{Centered triangular multivariate normal log-probability density function}
\usage{
centered_tri_multi_normal_lpdf(x, sigma = 1)
}
\arguments{
\item{x}{Matrix of quantiles.}

\item{sigma}{Standard deviation parameter(s).}
}
\value{
The log-density.
}
\description{
Centered triangular multivariate normal log-probability density function
}
