% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mcmc.R
\name{check_progress}
\alias{check_progress}
\title{Check MCMC progress from saved files}
\usage{
check_progress(name, dir = "BayesRTMB_mcmc", chains = 4)
}
\arguments{
\item{name}{Base name of the saved files.}

\item{dir}{Directory where progress files are saved. Default is "BayesRTMB_mcmc".}

\item{chains}{Number of chains to check.}
}
\description{
Check MCMC progress from saved files
}
