% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_post_estimation.R
\name{conditional_effects.mcmc_fit}
\alias{conditional_effects.mcmc_fit}
\title{Calculate conditional effects for MCMC fit objects}
\usage{
\method{conditional_effects}{mcmc_fit}(
  fit,
  effect,
  prob = 0.95,
  sd_multiplier = 1,
  resolution = 100,
  ...
)
}
\arguments{
\item{fit}{An object of class `MCMC_Fit`.}

\item{effect}{Name of the explanatory variable to visualize (e.g., "X1" or "X1:X2").}

\item{prob}{Probability for the credible/confidence interval (default is 0.95).}

\item{sd_multiplier}{Numeric. Multiplier for standard deviation when splitting continuous moderators (default is 1).}

\item{resolution}{Grid resolution to calculate for continuous variables (default is 100).}

\item{...}{Additional arguments.}
}
\value{
A `ce_rtmb` object.
}
\description{
Calculate conditional effects for MCMC fit objects
}
