% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{debate}
\alias{debate}
\title{Debate Simulation Data}
\format{
A data frame with 300 rows and 6 columns:
\describe{
  \item{group}{Group identifier. Each group consists of 3 members.}
  \item{sat}{Self-reported satisfaction level of the individual after the debate.}
  \item{talk}{Amount of talk or utterances by the individual during the debate.}
  \item{perf}{Evaluated performance of the group debate (group-level variable).}
  \item{skill}{Evaluated conversation skill of the individual.}
  \item{cond}{Experimental condition indicator (e.g., 0 = control, 1 = treatment).}
}
}
\source{
Simulated dummy data created by the package author.
}
\usage{
debate
}
\description{
A simulated dataset containing individual satisfaction scores and related
variables from a 3-person group debate experiment. This dummy data was
created by the package author to demonstrate hierarchical (multilevel) modeling.
}
\keyword{datasets}
