% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{Dim}
\alias{Dim}
\title{Define parameter dimensions and types}
\usage{
Dim(
  dim = 1,
  type = NULL,
  lower = NULL,
  upper = NULL,
  random = FALSE,
  names = NULL
)
}
\arguments{
\item{dim}{Dimensions of the parameter.}

\item{type}{Type of the parameter.}

\item{lower}{Lower bound.}

\item{upper}{Upper bound.}

\item{random}{Logical; whether it is a random effect.}

\item{names}{vector of parameters names}
}
\description{
Define parameter dimensions and types
}
