% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dirichlet_lpdf}
\alias{dirichlet_lpdf}
\title{Dirichlet log-probability density function}
\usage{
dirichlet_lpdf(x, alpha)
}
\arguments{
\item{x}{Vector or matrix of simplexes.}

\item{alpha}{Vector of concentration parameters.}
}
\value{
The sum of the log-density.
}
\description{
Dirichlet log-probability density function
}
\keyword{internal}
