% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satterthwaite.R
\name{.calculate_reml_satterthwaite_df_impl}
\alias{.calculate_reml_satterthwaite_df_impl}
\title{Calculate Satterthwaite degrees of freedom for REML models}
\usage{
.calculate_reml_satterthwaite_df_impl(model, ad_obj, opt_par, target_ran_idx)
}
\arguments{
\item{model}{An \code{RTMB_Model} object.}

\item{ad_obj}{An RTMB AD object.}

\item{opt_par}{Vector of optimized parameter values.}

\item{target_ran_idx}{Indices of target parameters within the random effect vector.}
}
\value{
A vector of estimated degrees of freedom for the target parameters.
}
\description{
Calculate Satterthwaite degrees of freedom for REML models
}
\keyword{internal}
