% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satterthwaite.R
\name{.calculate_satterthwaite_df_impl}
\alias{.calculate_satterthwaite_df_impl}
\title{Calculate Satterthwaite degrees of freedom for fixed effects}
\usage{
.calculate_satterthwaite_df_impl(
  model,
  ad_obj,
  idx_fix_active = NULL,
  L_u_total = NULL,
  opt_par = NULL
)
}
\arguments{
\item{model}{An \code{RTMB_Model} object.}

\item{ad_obj}{An RTMB AD object.}

\item{idx_fix_active}{Integer vector of active fixed parameter indices.}

\item{L_u_total}{Total number of unconstrained parameters.}

\item{opt_par}{Vector of optimized parameter values.}
}
\value{
A vector of estimated degrees of freedom.
}
\description{
Calculate Satterthwaite degrees of freedom for fixed effects
}
\keyword{internal}
