% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{.make_f_ad}
\alias{.make_f_ad}
\title{Factory function to create the internal AD objective function}
\usage{
.make_f_ad(data, par_list, log_prob, transform = NULL, jacobian_target = "all")
}
\arguments{
\item{data}{Model data.}

\item{par_list}{Parameter definitions.}

\item{log_prob}{User log-probability function.}

\item{transform}{User transform function.}

\item{jacobian_target}{Character; "all", "random", or "none".}
}
\value{
A function of one argument (unconstrained parameter list).
}
\description{
Creates a closure that can be passed to parallel workers
without carrying the entire R6 object.
}
\keyword{internal}
