% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model.R
\name{.null_model_impl}
\alias{.null_model_impl}
\title{Create a null model by fixing specified parameters}
\usage{
.null_model_impl(model, target, value = 0)
}
\arguments{
\item{model}{An \code{RTMB_Model} object.}

\item{target}{Character string specifying the target parameter and its prior.}

\item{value}{Numeric value to fix parameters to. Default is 0.}
}
\value{
A new \code{RTMB_Model} object with the specified parameters fixed.
}
\description{
Create a null model by fixing specified parameters
}
\keyword{internal}
