% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_helpers.R
\name{.print_code_impl}
\alias{.print_code_impl}
\title{Print model code and related information}
\usage{
.print_code_impl(model)
}
\arguments{
\item{model}{An \code{RTMB_Model} object.}
}
\description{
Print model code and related information
}
\keyword{internal}
