% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{env_to_ordered_list}
\alias{env_to_ordered_list}
\title{環境をリストに変換しつつ、元のリストの順序を保持する内部関数}
\usage{
env_to_ordered_list(env, orig_list, code_ast = NULL)
}
\description{
環境をリストに変換しつつ、元のリストの順序を保持する内部関数
}
\keyword{internal}
