% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_index.R
\name{ess_basic}
\alias{ess_basic}
\title{Basic Effective Sample Size for a single chain or pooled chains}
\usage{
ess_basic(sims)
}
\arguments{
\item{sims}{A numeric vector or matrix of samples.}
}
\value{
A numeric value.
}
\description{
Basic Effective Sample Size for a single chain or pooled chains
}
