% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_index.R
\name{ess_bulk}
\alias{ess_bulk}
\title{Calculate Bulk Effective Sample Size}
\usage{
ess_bulk(sims)
}
\arguments{
\item{sims}{A matrix of samples (iterations x chains).}
}
\value{
A numeric value.
}
\description{
Calculate Bulk Effective Sample Size
}
