% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_index.R
\name{ess_tail95}
\alias{ess_tail95}
\title{Calculate Tail Effective Sample Size (at 2.5\% and 97.5\% quantiles)}
\usage{
ess_tail95(sims)
}
\arguments{
\item{sims}{A matrix of samples (iterations x chains).}
}
\value{
A numeric value.
}
\description{
Calculate Tail Effective Sample Size (at 2.5\% and 97.5\% quantiles)
}
