% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{exponential_lpdf}
\alias{exponential_lpdf}
\title{Exponential log-probability density function}
\usage{
exponential_lpdf(x, rate)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{rate}{Vector of rates.}
}
\value{
The sum of the log-density.
}
\description{
Exponential log-probability density function
}
\keyword{internal}
