% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{fa_multi_normal_lpdf}
\alias{fa_multi_normal_lpdf}
\title{Factor analysis multivariate normal log-probability density function}
\usage{
fa_multi_normal_lpdf(x, mu, Lambda, psi)
}
\arguments{
\item{x}{Vector or matrix of quantiles.}

\item{mu}{Vector of means.}

\item{Lambda}{Factor loading matrix (P x K).}

\item{psi}{Vector of unique variances (P).}
}
\value{
The sum of the log-density.
}
\description{
Woodbury matrix identity is used for efficient computation.
}
