% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{fabs}
\alias{fabs}
\title{Smooth absolute value function}
\usage{
fabs(x, eps = 1e-08)
}
\arguments{
\item{x}{A numeric vector.}

\item{eps}{A small value added for numerical stability (default is 1e-8).}
}
\value{
The smooth absolute value of x.
}
\description{
Smooth absolute value function
}
