% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{fiml_multi_normal_CF_lpdf}
\alias{fiml_multi_normal_CF_lpdf}
\title{Multivariate normal log-probability density function parameterized by Cholesky factor of correlation matrix (FIML)}
\usage{
fiml_multi_normal_CF_lpdf(x, mean, sd, CF_Omega)
}
\arguments{
\item{x}{Vector or matrix of quantiles containing NAs.}

\item{mean}{Vector of means.}

\item{sd}{Vector of standard deviations.}

\item{CF_Omega}{Cholesky factor of the correlation matrix.}
}
\value{
The sum of the log-density using Full Information Maximum Likelihood.
}
\description{
Multivariate normal log-probability density function parameterized by Cholesky factor of correlation matrix (FIML)
}
\keyword{internal}
