% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{fiml_multi_normal_fa_lpdf}
\alias{fiml_multi_normal_fa_lpdf}
\title{Factor analysis multivariate normal log-probability density function (FIML)}
\usage{
fiml_multi_normal_fa_lpdf(x, mu, Lambda, psi)
}
\arguments{
\item{x}{Matrix of quantiles containing NAs.}

\item{mu}{Vector of means.}

\item{Lambda}{Factor loading matrix (P x K).}

\item{psi}{Vector of unique variances (P).}
}
\value{
The sum of the log-density.
}
\description{
Factor analysis multivariate normal log-probability density function (FIML)
}
\keyword{internal}
