% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{gamma_lpdf}
\alias{gamma_lpdf}
\title{Gamma log-probability density function}
\usage{
gamma_lpdf(x, shape, rate)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{shape}{Shape parameter.}

\item{rate}{Rate parameter.}
}
\value{
The sum of the log-density.
}
\description{
Gamma log-probability density function
}
\keyword{internal}
