% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{half_normal_lpdf}
\alias{half_normal_lpdf}
\title{Half-Normal log-probability density function}
\usage{
half_normal_lpdf(x, sd)
}
\arguments{
\item{x}{Vector of quantiles (must be non-negative).}

\item{sd}{Vector of standard deviations (scale parameter).}
}
\value{
The sum of the log-density.
}
\description{
Half-Normal log-probability density function
}
\keyword{internal}
