% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMH.R
\name{IMH_method}
\alias{IMH_method}
\title{Hybrid IMH and Preconditioned MALA sampling using Laplace approximation}
\usage{
IMH_method(
  model,
  sampling,
  warmup,
  chain,
  update_progress = NULL,
  df = 4,
  mix_ratio = 0.8,
  eps_mala = 0.1
)
}
\arguments{
\item{model}{An RTMB automatic differentiation objective object (`ad_obj`).}

\item{sampling}{Integer. Number of sampling iterations.}

\item{warmup}{Integer. Number of warmup iterations.}

\item{chain}{Integer. The ID of the current MCMC chain.}

\item{update_progress}{An optional callback function used to update the progress bar. Default is NULL.}

\item{df}{Numeric. Degrees of freedom for the multivariate t-distribution proposal in IMH. Default is 4.}

\item{mix_ratio}{Numeric. The probability of choosing the IMH step. Default is 0.8 (80% IMH, 20% MALA).}

\item{eps_mala}{Numeric. The step size parameter for the pMALA step. Default is 0.5.}
}
\value{
A list containing MCMC results.
}
\description{
This function performs a mixture of Independent Metropolis-Hastings (IMH)
and Preconditioned Metropolis-Adjusted Langevin Algorithm (pMALA).
The IMH step uses a multivariate t-distribution centered at the MAP estimate.
The pMALA step uses gradient information preconditioned by the inverse Hessian.
}
