% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{inject_namespace}
\alias{inject_namespace}
\title{ユーザーのコード(AST)を探索し、パッケージの関数に名前空間を自動付与する関数}
\usage{
inject_namespace(expr, pkg = "BayesRTMB")
}
\arguments{
\item{expr}{評価前のコード(AST)}

\item{pkg}{対象のパッケージ名}
}
\description{
ユーザーのコード(AST)を探索し、パッケージの関数に名前空間を自動付与する関数
}
\keyword{internal}
