% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{inverse_gamma_lpdf}
\alias{inverse_gamma_lpdf}
\title{Inverse-gamma log-probability density function}
\usage{
inverse_gamma_lpdf(x, shape, scale)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{shape}{Shape parameter.}

\item{scale}{Scale parameter.}
}
\value{
The sum of the log-density.
}
\description{
Inverse-gamma log-probability density function
}
\keyword{internal}
