% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mcmc.R
\name{item_curve}
\alias{item_curve}
\title{Calculate Item Response Curve / Category Response Curve}
\usage{
item_curve(x, ...)
}
\arguments{
\item{x}{An object of class}

\item{...}{Additional arguments.}
}
\description{
Calculate Item Response Curve / Category Response Curve
}
