% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mcmc.R
\name{item_info.RTMB_Fit_Base}
\alias{item_info.RTMB_Fit_Base}
\title{RTMB_Fit_Base の項目情報関数}
\usage{
\method{item_info}{RTMB_Fit_Base}(x, theta_seq = seq(-4, 4, length.out = 100), items = NULL, ...)
}
\arguments{
\item{x}{An object of class RTMB_Fit_Base.}

\item{theta_seq}{評価する特性値（能力値）のシーケンス。}

\item{items}{計算対象を特定の項目に絞る場合のインデックスまたは項目名（省略可能）。}

\item{...}{Additional arguments.}
}
\description{
RTMB_Fit_Base の項目情報関数
}
