% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{laplace_lpdf}
\alias{laplace_lpdf}
\title{Laplace log-probability density function}
\usage{
laplace_lpdf(x, location = 0, scale = 1)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{location}{Location parameter.}

\item{scale}{Scale parameter.}
}
\value{
The sum of the log-density.
}
\description{
Laplace log-probability density function
}
\keyword{internal}
