% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{lkj_CF_corr_lpdf}
\alias{lkj_CF_corr_lpdf}
\title{LKJ correlation log-probability density function for Cholesky factors}
\usage{
lkj_CF_corr_lpdf(CF_Omega, eta = 1)
}
\arguments{
\item{CF_Omega}{Cholesky factor of a correlation matrix.}

\item{eta}{Shape parameter.}
}
\value{
The log-density.
}
\description{
LKJ correlation log-probability density function for Cholesky factors
}
