% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{lkj_corr_lpdf}
\alias{lkj_corr_lpdf}
\title{LKJ correlation log-probability density function}
\usage{
lkj_corr_lpdf(Omega, eta = 1)
}
\arguments{
\item{Omega}{Correlation matrix.}

\item{eta}{Shape parameter.}
}
\value{
The log-density.
}
\description{
LKJ correlation log-probability density function
}
\keyword{internal}
