% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{log1m_exp}
\alias{log1m_exp}
\title{Log of one minus exponential of x}
\usage{
log1m_exp(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The natural logarithm of 1 - exp(x).
}
\description{
Log of one minus exponential of x
}
