% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{log1p_exp}
\alias{log1p_exp}
\title{Log of one plus exponential of x}
\usage{
log1p_exp(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The natural logarithm of 1 + exp(x).
}
\description{
Log of one plus exponential of x
}
