% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{log_det_chol}
\alias{log_det_chol}
\title{Log determinant of a Cholesky factor}
\usage{
log_det_chol(L)
}
\arguments{
\item{L}{A lower triangular Cholesky factor matrix.}
}
\value{
The log determinant of the corresponding covariance matrix.
}
\description{
Log determinant of a Cholesky factor
}
