% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{log_mix}
\alias{log_mix}
\title{Log mixture of two probabilities}
\usage{
log_mix(theta, lp1, lp2)
}
\arguments{
\item{theta}{Mixing proportion (between 0 and 1).}

\item{lp1}{Log-probability of the first component.}

\item{lp2}{Log-probability of the second component.}
}
\value{
The log-probability of the mixture.
}
\description{
Log mixture of two probabilities
}
