% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{log_softmax}
\alias{log_softmax}
\title{Log-softmax function}
\usage{
log_softmax(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector of log-softmax probabilities.
}
\description{
Log-softmax function
}
