% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{Log-sum-exp function}
\usage{
log_sum_exp(x, y = NULL)
}
\arguments{
\item{x}{A numeric vector or scalar.}

\item{y}{An optional numeric scalar.}
}
\value{
The log of the sum of the exponentials.
}
\description{
Log-sum-exp function
}
