% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{log_sum_exp_matrix}
\alias{log_sum_exp_matrix}
\title{Log-sum-exp function for matrices (row-wise)}
\usage{
log_sum_exp_matrix(M)
}
\arguments{
\item{M}{A numeric matrix or advector matrix.}
}
\value{
A numeric vector of row-wise log-sum-exp.
}
\description{
Log-sum-exp function for matrices (row-wise)
}
