% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{logit_normal_lpdf}
\alias{logit_normal_lpdf}
\title{Logit-Normal log-probability density function}
\usage{
logit_normal_lpdf(x, mu, sd)
}
\arguments{
\item{x}{Vector of quantiles (must be strictly between 0 and 1).}

\item{mu}{Vector of means on the logit scale.}

\item{sd}{Vector of standard deviations on the logit scale.}
}
\value{
The sum of the log-density.
}
\description{
Logit-Normal log-probability density function
}
\keyword{internal}
