% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{lognormal_lpdf}
\alias{lognormal_lpdf}
\title{Lognormal log-probability density function}
\usage{
lognormal_lpdf(x, meanlog, sdlog)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{meanlog}{Mean of the distribution on the log scale.}

\item{sdlog}{Standard deviation of the distribution on the log scale.}
}
\value{
The sum of the log-density.
}
\description{
Lognormal log-probability density function
}
\keyword{internal}
