% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{lower_tri_normal_lpdf}
\alias{lower_tri_normal_lpdf}
\title{Lower-triangular normal log-probability density function}
\usage{
lower_tri_normal_lpdf(x, mean = 0, sd = 1)
}
\arguments{
\item{x}{A matrix of lower-triangular parameters.}

\item{mean}{Mean of the normal distribution.}

\item{sd}{Standard deviation of the normal distribution.}
}
\value{
The log-density calculated only for the non-zero lower triangular elements.
}
\description{
Lower-triangular normal log-probability density function
}
