% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpdf.R
\docType{data}
\name{lpdf}
\alias{lpdf}
\title{Log-density and log-mass functions}
\format{
An object of class \code{environment} of length 28.
}
\usage{
lpdf
}
\description{
`lpdf` is an environment containing log-density / log-mass functions
used to build `log_prob()` functions.
}
\keyword{datasets}
