% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_index.R
\name{map_est}
\alias{map_est}
\title{Maximum A Posteriori (MAP) Estimate}
\usage{
map_est(z)
}
\arguments{
\item{z}{A numeric vector of samples.}
}
\value{
A numeric value.
}
\description{
Maximum A Posteriori (MAP) Estimate
}
