% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\docType{data}
\name{math}
\alias{math}
\title{Mathematical helper functions}
\format{
An object of class \code{environment} of length 15.
}
\usage{
math
}
\description{
`math` is an environment containing mathematical helper functions
used to build `log_prob()` functions.
}
\keyword{datasets}
