% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{mixture_lpdf}
\alias{mixture_lpdf}
\title{Generic Mixture log-probability density function}
\usage{
mixture_lpdf(x, pi_w, lpdf_list, sum = TRUE)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{pi_w}{Vector or matrix of mixing proportions.}

\item{lpdf_list}{A list of log-density vectors (one for each cluster).}
}
\value{
The sum of the log-density.
}
\description{
Generic Mixture log-probability density function
}
\keyword{internal}
