% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_code}
\alias{model_code}
\title{Model Code Wrapper for RTMB}
\usage{
model_code(expr, env = parent.frame())
}
\arguments{
\item{expr}{A block of code containing model description.}

\item{env}{Environment to assign to the generated function.}
}
\value{
A standard R function object taking (dat, par).
}
\description{
Model Code Wrapper for RTMB
}
