% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{multi_cauchy_lpdf}
\alias{multi_cauchy_lpdf}
\title{Multivariate Cauchy log-probability density function}
\usage{
multi_cauchy_lpdf(x, mean, Sigma, sum = TRUE)
}
\arguments{
\item{x}{Vector or matrix of quantiles.}

\item{mean}{Vector or matrix of means.}

\item{Sigma}{Scale matrix.}
}
\value{
The sum of the log-density.
}
\description{
Multivariate Cauchy log-probability density function
}
\keyword{internal}
