% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{multi_normal_CF_lpdf}
\alias{multi_normal_CF_lpdf}
\title{Multivariate normal log-probability density function parameterized by Cholesky factor of correlation matrix}
\usage{
multi_normal_CF_lpdf(x, mean, sd, CF_Omega)
}
\arguments{
\item{x}{Vector or matrix of quantiles.}

\item{mean}{Vector or matrix of means.}

\item{sd}{Vector of standard deviations.}

\item{CF_Omega}{Cholesky factor of the correlation matrix.}
}
\value{
The sum of the log-density.
}
\description{
Multivariate normal log-probability density function parameterized by Cholesky factor of correlation matrix
}
