% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{multi_normal_lpdf}
\alias{multi_normal_lpdf}
\title{Multivariate normal log-probability density function}
\usage{
multi_normal_lpdf(x, mean, Sigma)
}
\arguments{
\item{x}{Vector or matrix of quantiles.}

\item{mean}{Vector or matrix of means.}

\item{Sigma}{Covariance matrix.}
}
\value{
The sum of the log-density.
}
\description{
Multivariate normal log-probability density function
}
\keyword{internal}
