% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{multi_student_t_lpdf}
\alias{multi_student_t_lpdf}
\title{Multivariate Student-t log-probability density function}
\usage{
multi_student_t_lpdf(x, df, mean, Sigma, sum = TRUE)
}
\arguments{
\item{x}{Vector or matrix of quantiles.}

\item{df}{Degrees of freedom.}

\item{mean}{Vector or matrix of means.}

\item{Sigma}{Scale matrix.}
}
\value{
The sum of the log-density.
}
\description{
Multivariate Student-t log-probability density function
}
\keyword{internal}
