% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{multinomial_lpmf}
\alias{multinomial_lpmf}
\title{Multinomial log-probability mass function}
\usage{
multinomial_lpmf(x, size, prob)
}
\arguments{
\item{x}{Vector of counts.}

\item{size}{Total number of trials.}

\item{prob}{Vector of probabilities for each category.}
}
\value{
The sum of the log-probability.
}
\description{
Multinomial log-probability mass function
}
