% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{neg_binomial_2_lpmf}
\alias{neg_binomial_2_lpmf}
\title{Negative binomial log-probability mass function (alternative parameterization)}
\usage{
neg_binomial_2_lpmf(x, mu, size)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{mu}{Mean parameter.}

\item{size}{Dispersion parameter.}
}
\value{
The sum of the log-probability.
}
\description{
Negative binomial log-probability mass function (alternative parameterization)
}
