% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{neg_binomial_lpmf}
\alias{neg_binomial_lpmf}
\title{Negative binomial log-probability mass function}
\usage{
neg_binomial_lpmf(x, size, prob)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{size}{Target for number of successful trials.}

\item{prob}{Probability of success in each trial.}
}
\value{
The sum of the log-probability.
}
\description{
Negative binomial log-probability mass function
}
\keyword{internal}
