% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{normal_lpdf}
\alias{normal_lpdf}
\title{Normal log-probability density function}
\usage{
normal_lpdf(x, mean, sd, sum = TRUE)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{mean}{Vector of means.}

\item{sd}{Vector of standard deviations.}

\item{sum}{Logical; if TRUE (default), returns the sum of log-densities. If FALSE, returns element-wise log-densities.}
}
\value{
The sum of the log-density (scalar) or a vector of log-densities.
}
\description{
Normal log-probability density function
}
\keyword{internal}
