% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{normal_mixture_lpdf}
\alias{normal_mixture_lpdf}
\title{Normal Mixture log-probability density function}
\usage{
normal_mixture_lpdf(x, pi_w, mean, sd)
}
\arguments{
\item{x}{Vector of quantiles (observations).}

\item{pi_w}{Vector of mixing proportions for each cluster.}

\item{mean}{Vector of means for each cluster.}

\item{sd}{Vector of standard deviations for each cluster.}
}
\value{
The sum of the log-density.
}
\description{
Normal Mixture log-probability density function
}
\keyword{internal}
