% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{ordered_logistic_lpmf}
\alias{ordered_logistic_lpmf}
\title{Ordered logistic log-probability mass function}
\usage{
ordered_logistic_lpmf(x, eta, cutpoints)
}
\arguments{
\item{x}{Vector of ordered categorical outcomes.}

\item{eta}{Linear predictor.}

\item{cutpoints}{Vector of cutpoints.}
}
\value{
The sum of the log-probability.
}
\description{
Ordered logistic log-probability mass function
}
