% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{parameters_code}
\alias{parameters_code}
\title{パラメータ定義用コードブロック}
\usage{
parameters_code(...)
}
\arguments{
\item{...}{`名前 = Dim(...)` の形式で指定されるパラメータ群}
}
\value{
評価が遅延されたパラメータのリスト (alist)
}
\description{
`rtmb_model` に渡すパラメータのリストを定義します。
変数の評価をモデル構築時まで遅延させることで、次元指定（Dim）の厳密なエラーチェックを可能にします。
}
