% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_post_estimation.R
\name{plot.ce_rtmb}
\alias{plot.ce_rtmb}
\title{Plot method for ce_rtmb class (Base R)}
\usage{
\method{plot}{ce_rtmb}(x, ...)
}
\arguments{
\item{x}{An object of class ce_rtmb}

\item{...}{Additional arguments.}
}
\description{
Plot method for ce_rtmb class (Base R)
}
