% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_means.R
\name{plot.rtmb_lsmeans}
\alias{plot.rtmb_lsmeans}
\title{Plot marginal means with error bars}
\usage{
\method{plot}{rtmb_lsmeans}(
  x,
  y,
  error_bar = "se",
  col = NULL,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{x}{An rtmb_lsmeans object.}

\item{y}{Ignored.}

\item{error_bar}{Type of error bar ("se" or "ci").}

\item{col}{Color of the bars.}

\item{main}{Plot title.}

\item{xlab}{X axis label.}

\item{ylab}{Y axis label.}

\item{...}{Additional arguments passed to barplot.}
}
\description{
Plot method for rtmb_lsmeans objects.
}
