% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_acf}
\alias{plot_acf}
\title{Plot autocorrelation for one variable across chains}
\usage{
plot_acf(x, var_idx = 1)
}
\arguments{
\item{x}{A 3D array of posterior samples with dimensions
`(iterations, chains, variables)`.}

\item{var_idx}{Integer. Index of the variable to plot.}
}
\value{
No return value. This function is called for its side effect
  of plotting.
}
\description{
Draw autocorrelation plots for a selected parameter across all chains.
}
