% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_item_curve}
\alias{plot_item_curve}
\title{Plot item/category response curves}
\usage{
plot_item_curve(fit, ...)
}
\arguments{
\item{fit}{A fitted IRT model object.}

\item{...}{Additional arguments passed to \code{\link{item_curve}} or the plot method.}
}
\value{
Invisibly returns the plotted object of class \code{rtmb_item_curve}.
}
\description{
Calculate and plot item/category response curves for a fitted IRT model.
}
