% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_item_info}
\alias{plot_item_info}
\title{Plot item information functions}
\usage{
plot_item_info(fit, ...)
}
\arguments{
\item{fit}{A fitted IRT model object.}

\item{...}{Additional arguments passed to \code{\link{item_info}} or the plot method.}
}
\value{
Invisibly returns the plotted object of class \code{rtmb_item_info}.
}
\description{
Calculate and plot item information functions for a fitted IRT model.
}
