% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_lsmeans}
\alias{plot_lsmeans}
\title{Plot least-squares marginal means}
\usage{
plot_lsmeans(fit, specs, ...)
}
\arguments{
\item{fit}{A fitted model object (e.g. `Classic_Fit`).}

\item{specs}{Character vector specifying the variables for which marginal means are calculated.}

\item{...}{Additional arguments passed to \code{\link{lsmeans}} or the plot method.}
}
\value{
Invisibly returns the plotted object.
}
\description{
Calculate and plot marginal means for a fitted model.
}
