% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_means.R
\name{plot_means}
\alias{plot_means}
\title{Plot marginal means with error bars}
\usage{
plot_means(
  object,
  error_bar = "se",
  col = NULL,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{object}{An rtmb_lsmeans object.}

\item{error_bar}{Type of error bar.}

\item{col}{Color of the bars.}

\item{main}{Plot title.}

\item{xlab}{X axis label.}

\item{ylab}{Y axis label.}

\item{...}{Additional arguments.}
}
\description{
plot_means function creates a bar plot of marginal means with error bars.
}
