% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_test_info}
\alias{plot_test_info}
\title{Plot test information function}
\usage{
plot_test_info(fit, ...)
}
\arguments{
\item{fit}{A fitted IRT model object.}

\item{...}{Additional arguments passed to \code{\link{test_info}} or the plot method.}
}
\value{
Invisibly returns the plotted object of class \code{rtmb_test_info}.
}
\description{
Calculate and plot test information function for a fitted IRT model.
}
