% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{poisson_lpmf}
\alias{poisson_lpmf}
\title{Poisson log-probability mass function}
\usage{
poisson_lpmf(x, mean)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{mean}{Expected value.}
}
\value{
The sum of the log-probability.
}
\description{
Poisson log-probability mass function
}
