% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{positive_centered_tri_mvnormal_lpdf}
\alias{positive_centered_tri_mvnormal_lpdf}
\title{Positive centered triangular multivariate normal log-probability density function}
\usage{
positive_centered_tri_mvnormal_lpdf(x, sigma = 1)
}
\arguments{
\item{x}{Matrix of quantiles (centered triangular matrix with positive diagonals).}

\item{sigma}{Standard deviation parameter(s).}
}
\value{
The log-density.
}
\description{
Positive centered triangular multivariate normal log-probability density function
}
