% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_post_estimation.R
\name{print.bayes_factor_rtmb}
\alias{print.bayes_factor_rtmb}
\title{Print method for bayes_factor_rtmb objects}
\usage{
\method{print}{bayes_factor_rtmb}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class bayes_factor_rtmb.}

\item{digits}{Number of decimal places to display (default is 4).}

\item{...}{Additional arguments.}
}
\description{
Print method for bayes_factor_rtmb objects
}
