% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_post_estimation.R
\name{print.ce_simple}
\alias{print.ce_simple}
\title{Print simple effects}
\usage{
\method{print}{ce_simple}(x, digits = 3, ...)
}
\arguments{
\item{x}{A `ce_simple` object.}

\item{digits}{Number of digits to print.}

\item{...}{Additional arguments.}
}
\description{
Print simple effects
}
