% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mcmc.R
\name{print.summary_BayesRTMB}
\alias{print.summary_BayesRTMB}
\title{print for summary_BayesRTMB class}
\usage{
\method{print}{summary_BayesRTMB}(x, digits = 2, ...)
}
\arguments{
\item{digits}{integer}
}
\description{
print for summary_BayesRTMB class
}
