% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_prior.R
\name{prior_flat}
\alias{prior_flat}
\title{Specify a flat prior}
\usage{
prior_flat()
}
\value{
A list with class `"rtmb_prior"` and `type = "flat"`.
}
\description{
`prior_flat()` specifies that no additional prior density is added by the
wrapper. This is the only prior type allowed for `classic()`.
}
