% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_prior.R
\name{prior_rhs}
\alias{prior_rhs}
\title{Specify a Regularized Horseshoe prior for continuous shrinkage}
\usage{
prior_rhs(expected_vars = 3, slab_scale = 2, slab_df = 4, ...)
}
\arguments{
\item{expected_vars}{Expected number of non-zero variables. Default is 3.}

\item{slab_scale}{Scale parameter for the slab distribution. Default is 2.0.}

\item{slab_df}{Degrees of freedom for the slab distribution. Default is 4.0.}

\item{...}{Optional hyperparameters for other distributions.}
}
\value{
A list with class "rtmb_prior"
}
\description{
Specify a Regularized Horseshoe prior for continuous shrinkage
}
