% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_prior.R
\name{prior_ssp}
\alias{prior_ssp}
\title{Specify a Spike-and-Slab prior for variable selection}
\usage{
prior_ssp(ssp_ratio = 0.25, max_beta = 1, ...)
}
\arguments{
\item{ssp_ratio}{Prior probability of inclusion for each variable. Default is 0.25.}

\item{max_beta}{Maximum expected effect size. Default is 1.0.}

\item{...}{Optional hyperparameters for other distributions.}
}
\value{
A list with class "rtmb_prior"
}
\description{
Specify a Spike-and-Slab prior for variable selection
}
