% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_prior.R
\name{prior_weak}
\alias{prior_weak}
\title{Specify a weakly informative prior}
\usage{
prior_weak(sd_ratio = 0.5, max_beta = 1, ...)
}
\arguments{
\item{sd_ratio}{Ratio of the prior standard deviation to the half-range of the response variable. Default is 0.5.}

\item{max_beta}{Maximum expected effect size. Default is 1.0.}

\item{...}{Optional hyperparameters for other distributions.}
}
\value{
A list with class "rtmb_prior"
}
\description{
Specify a weakly informative prior
}
