% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{quad_form_chol}
\alias{quad_form_chol}
\title{Quadratic form using a Cholesky factor}
\usage{
quad_form_chol(x, L)
}
\arguments{
\item{x}{A numeric vector.}

\item{L}{A lower triangular Cholesky factor matrix.}
}
\value{
The value of the quadratic form.
}
\description{
Quadratic form using a Cholesky factor
}
