% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{quad_form_diag}
\alias{quad_form_diag}
\title{Quadratic form with a diagonal matrix}
\usage{
quad_form_diag(m, v)
}
\arguments{
\item{m}{A numeric matrix.}

\item{v}{A numeric vector representing the diagonal elements.}
}
\value{
The value of the quadratic form.
}
\description{
Quadratic form with a diagonal matrix
}
