% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_index.R
\name{quantile95}
\alias{quantile95}
\title{Calculate 95\% Quantiles}
\usage{
quantile95(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector of length 2.
}
\description{
Calculate 95\% Quantiles
}
