% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_index.R
\name{r_hat}
\alias{r_hat}
\title{Calculate Rank-normalized Split-R-hat}
\usage{
r_hat(sims)
}
\arguments{
\item{sims}{A matrix of samples (iterations x chains).}
}
\value{
A numeric value.
}
\description{
Calculate Rank-normalized Split-R-hat
}
