% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mcmc.R
\name{read_mcmc_csv}
\alias{read_mcmc_csv}
\title{Restore MCMC Fit from CSV}
\usage{
read_mcmc_csv(model, name, dir = "BayesRTMB_mcmc", chains = 4, laplace = FALSE)
}
\arguments{
\item{model}{An RTMB_Model object.}

\item{name}{Base name of the saved CSVs.}

\item{dir}{Directory where CSVs are saved. Default is "BayesRTMB_mcmc".}

\item{chains}{Number of chains. Default is 4.}

\item{laplace}{Logical; whether Laplace approximation was used. Default is FALSE.}
}
\value{
An MCMC_Fit object.
}
\description{
Restore MCMC Fit from CSV
}
