% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{register_lpdf}
\alias{register_lpdf}
\title{Register a custom likelihood function to the lpdf environment}
\usage{
register_lpdf(name, fun, force = FALSE)
}
\arguments{
\item{name}{The name of the likelihood function to register (character string).}

\item{fun}{The function to calculate the likelihood.}

\item{force}{Logical; whether to overwrite an existing function with the same name (default is FALSE).}
}
\description{
Register a custom likelihood function to the lpdf environment
}
