% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{register_math}
\alias{register_math}
\title{カスタム数学関数を math 環境に登録する}
\usage{
register_math(name, fun, force = FALSE)
}
\arguments{
\item{name}{登録する数学関数の名前（文字列）}

\item{fun}{登録する関数}

\item{force}{すでに同名の関数が存在する場合に上書きするかどうか（デフォルトはFALSE）}
}
\description{
カスタム数学関数を math 環境に登録する
}
