% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{register_obs}
\alias{register_obs}
\title{Internal function to register observations during likelihood calculation}
\usage{
register_obs(ll)
}
\description{
Internal function to register observations during likelihood calculation
}
\keyword{internal}
