% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model.R
\name{remove_prior_from_ast}
\alias{remove_prior_from_ast}
\title{Internal function to remove a prior from AST}
\usage{
remove_prior_from_ast(expr, target_spec, p_names = NULL)
}
\description{
Internal function to remove a prior from AST
}
\keyword{internal}
