% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{rtmb_code}
\alias{rtmb_code}
\title{StanライクなRTMBモデル定義ブロック (ハイブリッド対応版)}
\usage{
rtmb_code(...)
}
\arguments{
\item{...}{モデル定義ブロック}
}
\value{
未評価のコードブロックのリスト
}
\description{
R標準のカンマ区切り記法と、Stan風のカンマなし記法({}の内側での記述)の両方に対応します。
}
