% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{rtmb_hs_lm}
\alias{rtmb_hs_lm}
\title{RTMB-based Regularized Horseshoe Linear Regression}
\usage{
rtmb_hs_lm(
  formula,
  data,
  expected_vars = 3,
  slab_scale = 2,
  slab_df = 4,
  min_y = NULL,
  max_y = NULL,
  prior = list(prior_ratio = 1)
)
}
\arguments{
\item{formula}{Formula (e.g., Y ~ X1 + X2)}

\item{data}{Data frame}

\item{expected_vars}{Prior expected number of truly effective variables}

\item{slab_scale}{Scale of the slab (tails). Default is 2.0.}

\item{slab_df}{Degrees of freedom of the slab. Default is 4.0.}

\item{min_y}{Theoretical minimum value of the response variable (calculated from data if omitted)}

\item{max_y}{Theoretical maximum value of the response variable (calculated from data if omitted)}

\item{prior}{List of hyperparameters for prior distributions}
}
\description{
RTMB-based Regularized Horseshoe Linear Regression
}
