% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{select_parameters}
\alias{select_parameters}
\title{Select parameters from a list by name or index}
\usage{
select_parameters(all_list, pars)
}
\arguments{
\item{all_list}{A named list of parameters.}

\item{pars}{A character or numeric vector.}
}
\value{
A subset of the list.
}
\description{
Select parameters from a list by name or index
}
\keyword{internal}
