% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{sequential_logistic_lpmf}
\alias{sequential_logistic_lpmf}
\title{Sequential logistic log-probability mass function}
\usage{
sequential_logistic_lpmf(x, eta, cutpoints, sum = TRUE)
}
\arguments{
\item{x}{Vector of ordered categorical outcomes.}

\item{eta}{Linear predictor.}

\item{cutpoints}{Vector of sequential cutpoints.}
}
\value{
The sum of the log-probability.
}
\description{
Sequential logistic log-probability mass function
}
\keyword{internal}
