% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_post_estimation.R
\name{simple_effects.mcmc_fit}
\alias{simple_effects.mcmc_fit}
\title{Simple effects for MCMC fit objects}
\usage{
\method{simple_effects}{mcmc_fit}(fit, effect, prob = 0.95, sd_multiplier = 1, ...)
}
\arguments{
\item{fit}{An object of class `MCMC_Fit`.}

\item{effect}{Interaction term (e.g., "A:B").}

\item{prob}{Probability for credible intervals.}

\item{sd_multiplier}{Multiplier for SD for continuous moderators.}

\item{...}{Additional arguments.}
}
\description{
Simple effects for MCMC fit objects
}
