% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{softmax}
\alias{softmax}
\title{Softmax function}
\usage{
softmax(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector of softmax probabilities.
}
\description{
Softmax function
}
