% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_post_estimation.R
\name{sort_loadings}
\alias{sort_loadings}
\title{Sort and display factor loadings neatly}
\usage{
sort_loadings(loadings, cutoff = 0, round_digits = 3)
}
\arguments{
\item{loadings}{Matrix, data frame, or list of factor loadings (if list, the first element is used)}

\item{cutoff}{Absolute loadings below this value will be displayed as blank (default is 0.0)}

\item{round_digits}{Number of decimal places to display (default is 3)}
}
\value{
Sorted loading matrix (returned invisibly, allowing assignment to a variable)
}
\description{
Sort and display factor loadings neatly
}
