% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{squared_distance}
\alias{squared_distance}
\title{Squared Euclidean distance}
\usage{
squared_distance(x, y, eps = 1e-08)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}

\item{eps}{A small value added for numerical stability (default is 1e-8).}
}
\value{
The squared Euclidean distance between x and y.
}
\description{
Squared Euclidean distance
}
