% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stations}
\alias{stations}
\title{Stations Data}
\format{
A data frame or matrix containing relational data between stations.
}
\usage{
stations
}
\description{
A dataset representing distances or other relational metrics between stations.
}
\keyword{datasets}
