% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{student_t_lpdf}
\alias{student_t_lpdf}
\title{Student-t log-probability density function}
\usage{
student_t_lpdf(x, df, mu = 0, sigma = 1)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{df}{Degrees of freedom.}

\item{mu}{Location parameter.}

\item{sigma}{Scale parameter.}
}
\value{
The sum of the log-density.
}
\description{
Student-t log-probability density function
}
\keyword{internal}
