% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{sufficient_multi_normal_lpdf}
\alias{sufficient_multi_normal_lpdf}
\title{Sufficient statistics multivariate normal log-probability density function (covariance parameterization)}
\usage{
sufficient_multi_normal_lpdf(S_mat, N, y_bar, mean, Sigma)
}
\arguments{
\item{S_mat}{Deviation sum of squares matrix.}

\item{N}{Sample size.}

\item{y_bar}{Sample mean vector.}

\item{mean}{Mean parameter vector.}

\item{Sigma}{Covariance matrix.}
}
\value{
The exact log-likelihood of the N raw observations.
}
\description{
Sufficient statistics multivariate normal log-probability density function (covariance parameterization)
}
\keyword{internal}
