% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{sufficient_mvnorm_CF_lpdf}
\alias{sufficient_mvnorm_CF_lpdf}
\title{Sufficient statistics multivariate normal log-probability density function}
\usage{
sufficient_mvnorm_CF_lpdf(S_mat, N, y_bar, mean, sd, CF_Omega)
}
\arguments{
\item{S_mat}{Deviation sum of squares matrix.}

\item{N}{Sample size.}

\item{y_bar}{Sample mean vector.}

\item{mean}{Mean parameter vector.}

\item{sd}{Standard deviation parameter vector.}

\item{CF_Omega}{Cholesky factor of correlation matrix.}
}
\value{
The exact log-likelihood of the N raw observations.
}
\description{
Sufficient statistics multivariate normal log-probability density function
}
