% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{sum_to_zero}
\alias{sum_to_zero}
\title{Sum-to-zero transformation}
\usage{
sum_to_zero(x)
}
\arguments{
\item{x}{A numeric vector of length K-1.}
}
\value{
A numeric vector of length K whose elements sum to zero.
}
\description{
Transforms a vector of length K-1 to a vector of length K that sums to zero
using an orthogonal basis matrix.
}
