% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{sum_to_zero_multi_norm_lpdf}
\alias{sum_to_zero_multi_norm_lpdf}
\title{Sum-to-zero multivariate normal log-probability density function}
\usage{
sum_to_zero_multi_norm_lpdf(x, sigma = 1, K = NULL)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{sigma}{Standard deviation parameter.}

\item{K}{Dimension. If NULL, inferred from the length of x.}
}
\value{
The log-density.
}
\description{
Sum-to-zero multivariate normal log-probability density function
}
