% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_post_estimation.R
\name{summary.ce_rtmb}
\alias{summary.ce_rtmb}
\title{Summary method for ce_rtmb class}
\usage{
\method{summary}{ce_rtmb}(object, ...)
}
\arguments{
\item{object}{An object of class ce_rtmb}

\item{...}{Additional arguments.}
}
\description{
Summary method for ce_rtmb class
}
