% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{to_centered_matrix}
\alias{to_centered_matrix}
\title{Vector to centered matrix (RTMB compatible)}
\usage{
to_centered_matrix(x, R, C)
}
\arguments{
\item{x}{A numeric vector of length (R-1) * C.}

\item{R}{Number of rows.}

\item{C}{Number of columns.}
}
\value{
An R x C matrix where each column sums to zero.
}
\description{
Vector to centered matrix (RTMB compatible)
}
