% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{to_centered_tri}
\alias{to_centered_tri}
\title{Vector to centered triangular matrix (RTMB compatible)}
\usage{
to_centered_tri(x, R, C)
}
\arguments{
\item{x}{A numeric vector of appropriate length.}

\item{R}{Number of rows.}

\item{C}{Number of columns.}
}
\value{
An R x C matrix with column-wise sum-to-zero constraints on lower elements.
}
\description{
Vector to centered triangular matrix (RTMB compatible)
}
