% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{to_lower_tri}
\alias{to_lower_tri}
\title{Vector to lower triangular matrix (RTMB compatible)}
\usage{
to_lower_tri(x, M, D)
}
\arguments{
\item{x}{A numeric or advector.}

\item{M}{Number of rows.}

\item{D}{Number of columns.}
}
\value{
An M x D lower triangular matrix.
}
\description{
Vector to lower triangular matrix (RTMB compatible)
}
