% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{transform_code}
\alias{transform_code}
\title{Transformed Code Wrapper for RTMB}
\usage{
transform_code(expr, env = parent.frame(), ad_seed_name = NULL)
}
\arguments{
\item{expr}{A block of code containing calculations for transformed parameters.}

\item{env}{Environment to assign to the generated function.}

\item{ad_seed_name}{Optional parameter name used internally to seed RTMB's AD type.}
}
\value{
A function taking (dat, par) that returns a named list.
}
\description{
Transformed Code Wrapper for RTMB
}
