% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{uniform_lpdf}
\alias{uniform_lpdf}
\title{Uniform log-probability density function}
\usage{
uniform_lpdf(x, a, b)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{a}{Lower limit of the distribution.}

\item{b}{Upper limit of the distribution.}
}
\value{
The sum of the log-density.
}
\description{
Uniform log-probability density function
}
