% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{weibull_lpdf}
\alias{weibull_lpdf}
\title{Weibull log-probability density function}
\usage{
weibull_lpdf(x, shape, scale)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{shape}{Shape parameter.}

\item{scale}{Scale parameter.}
}
\value{
The sum of the log-density.
}
\description{
Weibull log-probability density function
}
