% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{wishart_CF_lpdf}
\alias{wishart_CF_lpdf}
\title{Wishart log-probability density function parameterized by Cholesky factor of correlation matrix}
\usage{
wishart_CF_lpdf(X, n, sd, CF_Omega)
}
\arguments{
\item{X}{Symmetric positive-definite matrix (scatter matrix).}

\item{n}{Degrees of freedom (sample size N).}

\item{sd}{Vector of standard deviations.}

\item{CF_Omega}{Cholesky factor of the correlation matrix.}
}
\value{
The log-density with all constant terms.
}
\description{
Wishart log-probability density function parameterized by Cholesky factor of correlation matrix
}
\keyword{internal}
