% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{wishart_lpdf}
\alias{wishart_lpdf}
\title{Wishart log-probability density function}
\usage{
wishart_lpdf(X, n, V)
}
\arguments{
\item{X}{Symmetric positive-definite matrix (scatter matrix).}

\item{n}{Degrees of freedom (sample size N).}

\item{V}{Scale matrix (covariance matrix Omega).}
}
\value{
The log-density with all constant terms.
}
\description{
Wishart log-probability density function
}
