% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AVDI.R
\name{ADVI_method}
\alias{ADVI_method}
\title{Automatic Differentiation Variational Inference (ADVI)}
\usage{
ADVI_method(
  model,
  par_list,
  pl_full,
  iter = 10000,
  min_iter = 1000,
  tol_rel_obj = 0.01,
  window_size = 100,
  num_samples = 1000,
  chains = 1,
  alpha = 0.01,
  laplace = FALSE,
  print_freq = 500,
  fullrank = FALSE
)
}
\arguments{
\item{model}{An RTMB objective function object (`ad_obj`).}

\item{par_list}{A list defining the structure of parameters to be estimated.}

\item{pl_full}{A list defining the full structure of parameters including random effects.}

\item{iter}{Integer; maximum number of iterations for the optimization. Default is 10000.}

\item{min_iter}{Integer; minimum number of iterations to run before checking for convergence. Default is 1000.}

\item{tol_rel_obj}{Numeric; relative tolerance for the ELBO to determine convergence. Default is 0.01.}

\item{window_size}{Integer; size of the moving window to calculate the median ELBO for the convergence check. Default is 100.}

\item{num_samples}{Integer; number of posterior draws to generate after optimization. Default is 1000.}

\item{chains}{Integer; number of chains to structure the output array (for compatibility). Default is 1.}

\item{alpha}{Numeric; learning rate (step size) for the Adam optimizer. Default is 0.01.}

\item{laplace}{Logical; whether Laplace approximation is used. Default is FALSE.}

\item{print_freq}{Integer; frequency of printing progress to the console. Set to 0 to disable. Default is 500.}

\item{fullrank}{Logical; whether to use a full-rank multivariate normal distribution for the approximation. Default is FALSE.}
}
\value{
A list containing `fit`, `random_fit`, `elbo_history`, and `elbo_final`.
}
\description{
Automatic Differentiation Variational Inference (ADVI)
}
