% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AVDI.R
\name{ADVI_method}
\alias{ADVI_method}
\title{Automatic Differentiation Variational Inference (ADVI)}
\usage{
ADVI_method(
  model,
  par_list,
  pl_full,
  iter = 10000,
  tol_rel_obj = 0.01,
  window_size = 100,
  num_samples = 1000,
  chains = 1,
  alpha = 0.01,
  laplace = FALSE
)
}
\arguments{
\item{model}{RTMB ad_obj}

\item{par_list}{Parameter definitions}

\item{pl_full}{Full parameter definitions}

\item{iter}{Maximum number of iterations}

\item{tol_rel_obj}{Relative tolerance for convergence}

\item{window_size}{Window size for median smoothing}

\item{num_samples}{Number of posterior draws}

\item{chains}{Number of chains to format the output}

\item{alpha}{Learning rate for Adam}

\item{laplace}{Logical; whether Laplace approximation is used}
}
\description{
Automatic Differentiation Variational Inference (ADVI)
}
