% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_post_estimation.R
\name{bayes_factor}
\alias{bayes_factor}
\title{Calculate Bayes factor from log marginal likelihoods}
\usage{
bayes_factor(logml1, logml2)
}
\arguments{
\item{logml1}{Log marginal likelihood of Model 1 (e.g., target model)}

\item{logml2}{Log marginal likelihood of Model 2 (e.g., reference/null model)}
}
\value{
An object containing Bayes factor, log Bayes factor, estimation error, and interpretation
}
\description{
Calculate Bayes factor from log marginal likelihoods
}
\examples{
\dontrun{
  # Compare two models using Bayes Factor
  fit1 <- rtmb_lm(mpg ~ wt, data = mtcars)
  fit2 <- rtmb_lm(mpg ~ wt + hp, data = mtcars)
  mcmc1 <- fit1$sample(sampling = 500, warmup = 500)
  mcmc2 <- fit2$sample(sampling = 500, warmup = 500)
  bf <- bayes_factor(mcmc1, mcmc2)
  print(bf)
}
}
