% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mcmc.R
\name{bayes_factor}
\alias{bayes_factor}
\title{対数周辺尤度からベイズファクターを計算する関数}
\usage{
bayes_factor(logml1, logml2)
}
\arguments{
\item{logml1}{モデル1の対数周辺尤度 (対象モデルなど)}

\item{logml2}{モデル2の対数周辺尤度 (基準モデル/帰無モデルなど)}
}
\value{
ベイズファクター、対数ベイズファクター、推定誤差、および解釈を含むオブジェクト
}
\description{
対数周辺尤度からベイズファクターを計算する関数
}
