% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beverage}
\alias{beverage}
\title{Beverage Preference Data}
\format{
A data frame (or matrix) with [INSERT NUMBER OF ROWS] rows and [INSERT NUMBER OF COLUMNS] columns.
\describe{
  \item{Column1}{Description of the first column (e.g., preference score).}
  \item{Column2}{Description of the second column.}
  % Add other columns as necessary
}
}
\source{
Adachi, K. (2000). A Random Effect Model in Metric Multidimensional Unfolding.
\emph{The Japanese Journal of Behaviormetrics}, 27(1), 12-23.
\url{https://doi.org/10.2333/jbhmk.27.12}
}
\usage{
beverage
}
\description{
A dataset containing preference ratings for various beverages.
This data was used to demonstrate a random effect model in metric
multidimensional unfolding.
}
\keyword{datasets}
