% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{binomial_logit_lpmf}
\alias{binomial_logit_lpmf}
\title{Binomial log-probability mass function with logit parameterization}
\usage{
binomial_logit_lpmf(x, size, eta)
}
\arguments{
\item{x}{Vector of quantiles (number of successes).}

\item{size}{Number of trials.}

\item{eta}{Linear predictor (logit probability).}
}
\value{
The sum of the log-probability.
}
\description{
Binomial log-probability mass function with logit parameterization
}
\keyword{internal}
