% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classic.R
\name{Classic_Fit}
\alias{Classic_Fit}
\title{Classic fit object}
\description{
An R6 class representing the results of a classical (frequentist) estimation.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{The `Classic_Model` object used for estimation.}

\item{\code{fit}}{The result of the estimation (dataframe or lm object).}

\item{\code{par}}{A named list of parameter estimates.}

\item{\code{bootstrap_results}}{A matrix containing bootstrap samples, if applicable.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Classic_Fit-new}{\code{Classic_Fit$new()}}
\item \href{#method-Classic_Fit-summary}{\code{Classic_Fit$summary()}}
\item \href{#method-Classic_Fit-print}{\code{Classic_Fit$print()}}
\item \href{#method-Classic_Fit-.construct_par_list}{\code{Classic_Fit$.construct_par_list()}}
\item \href{#method-Classic_Fit-clone}{\code{Classic_Fit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Classic_Fit-new"></a>}}
\if{latex}{\out{\hypertarget{method-Classic_Fit-new}{}}}
\subsection{Method \code{new()}}{
Create a new `Classic_Fit` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Classic_Fit$new(model, fit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{The `Classic_Model` object.}

\item{\code{fit}}{The result of the estimation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Classic_Fit-summary"></a>}}
\if{latex}{\out{\hypertarget{method-Classic_Fit-summary}{}}}
\subsection{Method \code{summary()}}{
Display a summary of the estimation results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Classic_Fit$summary(digits = 5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{digits}}{Number of digits to print.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Classic_Fit-print"></a>}}
\if{latex}{\out{\hypertarget{method-Classic_Fit-print}{}}}
\subsection{Method \code{print()}}{
Print the fit results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Classic_Fit$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Classic_Fit-.construct_par_list"></a>}}
\if{latex}{\out{\hypertarget{method-Classic_Fit-.construct_par_list}{}}}
\subsection{Method \code{.construct_par_list()}}{
(Internal) Construct a list of parameters from the fit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Classic_Fit$.construct_par_list(fit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fit}}{The fit result (dataframe or lm object).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Classic_Fit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Classic_Fit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Classic_Fit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
