% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mcmc.R
\name{conditional_effects.mcmc_fit}
\alias{conditional_effects.mcmc_fit}
\title{MCMC_Fitクラス用のメソッド}
\usage{
\method{conditional_effects}{mcmc_fit}(fit, effect, resolution = 100, prob = 0.95, ...)

\method{conditional_effects}{mcmc_fit}(fit, effect, resolution = 100, prob = 0.95, ...)
}
\description{
MCMC_Fitクラス用のメソッド

ce_rtmb クラス専用のプロットメソッド (Base R)
}
