% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mcmc.R
\name{conditional_effects.mcmc_fit}
\alias{conditional_effects.mcmc_fit}
\title{Calculate conditional effects for MCMC fit objects}
\usage{
\method{conditional_effects}{mcmc_fit}(fit, effect, resolution = 100, prob = 0.95, ...)
}
\arguments{
\item{fit}{An object of class `MCMC_Fit`.}

\item{effect}{Name of the explanatory variable to visualize (e.g., "X1" or "X1:X2").}

\item{resolution}{Grid resolution to calculate for continuous variables (default is 100).}

\item{prob}{Probability for the credible interval (default is 0.95).}

\item{...}{Additional arguments.}
}
\description{
Calculate conditional effects for MCMC fit objects
}
