% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_post_estimation.R
\name{conditional_effects}
\alias{conditional_effects}
\title{Calculate Conditional Effects}
\usage{
conditional_effects(fit, effect, ...)
}
\arguments{
\item{fit}{Model fit object.}

\item{effect}{Name of the variable to visualize the effect.}

\item{...}{Additional arguments.}
}
\description{
Calculate Conditional Effects
}
\examples{
\dontrun{
  fit <- rtmb_lm(mpg ~ wt + hp, data = mtcars)
  mcmc_fit <- fit$sample(sampling = 500, warmup = 500)
  ce <- conditional_effects(mcmc_fit, effect = "wt")
  plot(ce)
}
}
