% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mcmc.R
\name{conditional_effects}
\alias{conditional_effects}
\title{Calculate Conditional Effects}
\usage{
conditional_effects(fit, effect, ...)
}
\arguments{
\item{fit}{Model fit object.}

\item{effect}{Name of the variable to visualize the effect.}

\item{...}{Additional arguments.}
}
\description{
Calculate Conditional Effects
}
