% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{Dim}
\alias{Dim}
\title{paramters functions}
\usage{
Dim(dim = 1, type = NULL, lower = NULL, upper = NULL, random = FALSE)
}
\arguments{
\item{dim}{パラメータの次元}

\item{type}{パラメータの型}

\item{lower}{下限}

\item{upper}{上限}

\item{random}{変量効果かどうか}
}
\description{
paramters functions
}
