% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{env_to_ordered_list}
\alias{env_to_ordered_list}
\title{Internal function to convert an environment to a list while maintaining original order}
\usage{
env_to_ordered_list(env, orig_list, code_ast = NULL)
}
\description{
Internal function to convert an environment to a list while maintaining original order
}
\keyword{internal}
