% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMH.R
\name{IMH_method}
\alias{IMH_method}
\title{Independent Metropolis-Hastings Algorithm}
\usage{
IMH_method(
  ad_obj,
  mu,
  Sigma,
  sampling,
  warmup,
  chain = 1,
  update_progress = NULL
)
}
\arguments{
\item{ad_obj}{AD object from RTMB}

\item{mu}{MAP estimates (mean vector)}

\item{Sigma}{Covariance matrix from Laplace approximation}

\item{sampling}{Number of sampling iterations}

\item{warmup}{Number of warmup iterations}

\item{chain}{Chain ID (for progress reporting)}

\item{update_progress}{Callback function for progress reporting}
}
\value{
A list containing samples and diagnostic statistics
}
\description{
Independent Metropolis-Hastings Algorithm
}
