% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{inject_namespace}
\alias{inject_namespace}
\title{Internal function to search AST and inject namespace to package functions}
\usage{
inject_namespace(expr, pkg = "BayesRTMB")
}
\arguments{
\item{expr}{Unevaluated code (AST)}

\item{pkg}{Target package name}
}
\description{
Internal function to search AST and inject namespace to package functions
}
\keyword{internal}
