% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mcmc.R
\name{item_curve.RTMB_Fit_Base}
\alias{item_curve.RTMB_Fit_Base}
\title{Item Response Curve for RTMB_Fit_Base}
\usage{
\method{item_curve}{RTMB_Fit_Base}(x, theta_seq = seq(-4, 4, length.out = 100), items = NULL, ...)
}
\arguments{
\item{x}{An object of class RTMB_Fit_Base.}

\item{theta_seq}{Sequence of trait values (ability) to evaluate.}

\item{items}{Index or item names to restrict the calculation to specific items (optional).}

\item{...}{Additional arguments.}
}
\description{
Item Response Curve for RTMB_Fit_Base
}
