% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_post_estimation.R
\name{item_info}
\alias{item_info}
\title{Calculate Item Information Function}
\usage{
item_info(x, ...)
}
\arguments{
\item{x}{An object of class RTMB_Fit_Base}

\item{...}{Additional arguments.}
}
\description{
Calculate Item Information Function
}
