% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classic.R
\name{lsmeans}
\alias{lsmeans}
\title{Least-squares means (marginal means)}
\usage{
lsmeans(object, specs, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{specs}{Variable names to calculate marginal means for.}

\item{...}{Additional arguments.}
}
\description{
Generic function to calculate least-squares means (marginal means) for fitted models.
}
