% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAP_Fit.R
\name{MAP_Fit}
\alias{MAP_Fit}
\title{MAP fit object}
\description{
MAP fit object

MAP fit object
}
\details{
An R6 class storing optimization results from maximum a posteriori
(MAP) estimation.
}
\section{Super class}{
\code{\link[BayesRTMB:RTMB_Fit_Base]{BayesRTMB::RTMB_Fit_Base}} -> \code{map_fit}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{The `RTMB_Model` object used for estimation.}

\item{\code{par_vec}}{Parameter vector on the unconstrained scale.}

\item{\code{par}}{Parameter list on the constrained scale.}

\item{\code{objective}}{RTMB objective function object.}

\item{\code{log_ml}}{Log marginal likelihood or related model criterion.}

\item{\code{convergence}}{Optimizer convergence code.}

\item{\code{sd_rep}}{Standard deviation report object.}

\item{\code{df_fixed}}{Summary table for fixed-effect parameters.}

\item{\code{random_effects}}{Random effect estimates.}

\item{\code{df_transform}}{Summary table for transformed parameter estimates.}

\item{\code{df_generate}}{Summary table for generated quantity estimates.}

\item{\code{opt_history}}{A vector of optimize objective history.}

\item{\code{transform}}{List of transformed parameters maintaining their original dimensions.}

\item{\code{generate}}{List of generated quantities maintaining their original dimensions.}

\item{\code{se_samples}}{List of simulated samples for standard error estimation.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-map_fit-get_point_estimate}{\code{MAP_Fit$get_point_estimate()}}
\item \href{#method-map_fit-EAP}{\code{MAP_Fit$EAP()}}
\item \href{#method-map_fit-MAP}{\code{MAP_Fit$MAP()}}
\item \href{#method-map_fit-new}{\code{MAP_Fit$new()}}
\item \href{#method-map_fit-summary}{\code{MAP_Fit$summary()}}
\item \href{#method-map_fit-print}{\code{MAP_Fit$print()}}
\item \href{#method-map_fit-generated_quantities}{\code{MAP_Fit$generated_quantities()}}
\item \href{#method-map_fit-clone}{\code{MAP_Fit$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="BayesRTMB" data-topic="RTMB_Fit_Base" data-id="fa_rotate"><a href='../../BayesRTMB/html/RTMB_Fit_Base.html#method-RTMB_Fit_Base-fa_rotate'><code>BayesRTMB::RTMB_Fit_Base$fa_rotate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="BayesRTMB" data-topic="RTMB_Fit_Base" data-id="rotate"><a href='../../BayesRTMB/html/RTMB_Fit_Base.html#method-RTMB_Fit_Base-rotate'><code>BayesRTMB::RTMB_Fit_Base$rotate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_fit-get_point_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-map_fit-get_point_estimate}{}}}
\subsection{Method \code{get_point_estimate()}}{
Get point estimate for a target parameter (internal use).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAP_Fit$get_point_estimate(target)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target}}{Target parameter name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Matrix or array of point estimate.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_fit-EAP"></a>}}
\if{latex}{\out{\hypertarget{method-map_fit-EAP}{}}}
\subsection{Method \code{EAP()}}{
Return point estimates (EAP is not applicable).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAP_Fit$EAP(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A named list of point estimates.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_fit-MAP"></a>}}
\if{latex}{\out{\hypertarget{method-map_fit-MAP}{}}}
\subsection{Method \code{MAP()}}{
Return point estimates (MAP sampling method is not applicable).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAP_Fit$MAP(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A named list of point estimates.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_fit-new"></a>}}
\if{latex}{\out{\hypertarget{method-map_fit-new}{}}}
\subsection{Method \code{new()}}{
Create a new `MAP_Fit` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAP_Fit$new(
  model,
  par_vec,
  par,
  objective,
  log_ml,
  convergence,
  sd_rep,
  df_fixed,
  random_effects,
  df_transform = NULL,
  df_generate = NULL,
  opt_history = NULL,
  transform = NULL,
  generate = NULL,
  se_samples = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{The `RTMB_Model` object used for estimation.}

\item{\code{par_vec}}{Parameter vector on the unconstrained scale.}

\item{\code{par}}{Parameter list on the constrained scale.}

\item{\code{objective}}{RTMB objective function object.}

\item{\code{log_ml}}{Log marginal likelihood or related model criterion.}

\item{\code{convergence}}{Optimizer convergence code.}

\item{\code{sd_rep}}{Standard deviation report object.}

\item{\code{df_fixed}}{Summary table for fixed-effect parameters.}

\item{\code{random_effects}}{Random effect estimates.}

\item{\code{df_transform}}{Summary table for transformed parameter estimates.}

\item{\code{df_generate}}{Summary table for generated quantity estimates.}

\item{\code{opt_history}}{A vector of optimize objective history.}

\item{\code{transform}}{List of transformed parameters maintaining their original dimensions.}

\item{\code{generate}}{List of generated quantities maintaining their original dimensions.}

\item{\code{se_samples}}{List of simulated samples for standard error estimation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_fit-summary"></a>}}
\if{latex}{\out{\hypertarget{method-map_fit-summary}{}}}
\subsection{Method \code{summary()}}{
Summarize MAP estimates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAP_Fit$summary(pars = NULL, max_rows = 10, digits = 5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pars}}{Character vector specifying the names of parameters to summarize. If NULL, all available parameters are summarized.}

\item{\code{max_rows}}{Maximum number of rows to print in summaries. Default is 10.}

\item{\code{digits}}{Number of digits to print.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A summary object, typically a data frame.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_fit-print"></a>}}
\if{latex}{\out{\hypertarget{method-map_fit-print}{}}}
\subsection{Method \code{print()}}{
Print a brief summary of the fitted object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAP_Fit$print(pars = NULL, max_rows = 10, digits = 5, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pars}}{Character vector specifying the names of parameters to summarize.}

\item{\code{max_rows}}{Maximum number of rows to print in summaries.}

\item{\code{digits}}{Number of digits to print.}

\item{\code{...}}{Additional arguments passed to the `summary` method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The object itself, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_fit-generated_quantities"></a>}}
\if{latex}{\out{\hypertarget{method-map_fit-generated_quantities}{}}}
\subsection{Method \code{generated_quantities()}}{
Compute generated quantities from the MAP estimate.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAP_Fit$generated_quantities(code)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{An `rtmb_code(\{ ... \})` or `\{ ... \}` block containing the logic to be calculated using the MAP estimate.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The `MAP_Fit` object itself (invisibly). Results are added or updated in the `generate` list and `df_generate`.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_fit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-map_fit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAP_Fit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
