% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{normal_lpdf}
\alias{normal_lpdf}
\title{Normal log-probability density function}
\usage{
normal_lpdf(x, mean, sd)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{mean}{Vector of means.}

\item{sd}{Vector of standard deviations.}
}
\value{
The sum of the log-density.
}
\description{
Normal log-probability density function
}
\keyword{internal}
