% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{parameters_code}
\alias{parameters_code}
\title{Code block for parameter definitions}
\usage{
parameters_code(expr)
}
\arguments{
\item{expr}{A parameter definition expression enclosed in `{}`.}
}
\value{
A lazily evaluated list of parameters.
}
\description{
Defines the list of parameters to be passed to `rtmb_model`.
By writing in the format `variable_name = Dim(...)` within a block enclosed in `{}`,
evaluation is delayed, enabling strict error checking during model construction.
}
