% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{parameters_code}
\alias{parameters_code}
\title{パラメータ定義用コードブロック}
\usage{
parameters_code(expr)
}
\arguments{
\item{expr}{`{}` で囲まれたパラメータ定義の式}
}
\value{
評価が遅延されたパラメータのリスト
}
\description{
`rtmb_model` に渡すパラメータのリストを定義します。
`{}` で囲まれたブロック内で `変数名 = Dim(...)` の形式で記述することで、
評価を遅延させ、モデル構築時の厳密なエラーチェックを可能にします。
}
