% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_forest}
\alias{plot_forest}
\title{Plot parameter estimates and credible intervals (Forest Plot)}
\usage{
plot_forest(x, prob = 0.95)
}
\arguments{
\item{x}{A 3D array of posterior samples with dimensions `(iterations, chains, variables)`.}

\item{prob}{Numeric. Probability mass for the credible interval (default is 0.95).}
}
\value{
No return value.
}
\description{
Plot parameter estimates and credible intervals (Forest Plot)
}
