% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{positive_lower_tri_normal_lpdf}
\alias{positive_lower_tri_normal_lpdf}
\title{Positive centered triangular multivariate normal log-probability density function}
\usage{
positive_lower_tri_normal_lpdf(x, mean = 0, sd = 1)
}
\arguments{
\item{x}{A matrix of lower-triangular parameters (Cholesky factor with positive diagonals).}

\item{mean}{Mean of the normal distribution (assumed to be 0 for the half-normal correction).}

\item{sd}{Standard deviation of the normal distribution.}

\item{sigma}{Standard deviation parameter(s).}
}
\value{
The log-density.

The log-density calculated for the non-zero lower triangular elements.
}
\description{
Positive centered triangular multivariate normal log-probability density function
}
