% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_post_estimation.R
\name{print.bayes_factor}
\alias{print.bayes_factor}
\title{Print method for bayes_factor objects}
\usage{
\method{print}{bayes_factor}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class bayes_factor.}

\item{digits}{Number of decimal places to display (default is 4).}

\item{...}{Additional arguments.}
}
\description{
Print method for bayes_factor objects
}
