% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mcmc.R
\name{print.ce_rtmb}
\alias{print.ce_rtmb}
\title{ce_rtmb クラスのprintメソッド (自動的にplotを呼ぶ)}
\usage{
\method{print}{ce_rtmb}(x, ...)
}
\arguments{
\item{x}{An object of class ce_rtmb}

\item{...}{Additional arguments.}
}
\description{
ce_rtmb クラスのprintメソッド (自動的にplotを呼ぶ)
}
