% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mcmc.R
\name{print.summary_BayesRTMB}
\alias{print.summary_BayesRTMB}
\title{summary_BayesRTMB クラス専用の print メソッド}
\usage{
\method{print}{summary_BayesRTMB}(x, ...)
}
\description{
summary_BayesRTMB クラス専用の print メソッド
}
