% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_prior.R
\name{prior_jzs}
\alias{prior_jzs}
\title{Specify a JZS (Jeffrey-Zellner-Siow) prior for t-tests}
\usage{
prior_jzs(r = 0.707)
}
\arguments{
\item{r}{Numeric; scale parameter for the Cauchy prior on effect size delta. Default is 0.707.}
}
\value{
A list with class "rtmb_prior"
}
\description{
Specify a JZS (Jeffrey-Zellner-Siow) prior for t-tests
}
