% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{prior_uniform}
\alias{prior_uniform}
\title{Specify a uniform or manual prior}
\usage{
prior_uniform(
  Intercept_sd = NULL,
  b_sd = NULL,
  mu_sd = NULL,
  sigma_rate = NULL,
  tau_rate = NULL,
  ...
)
}
\arguments{
\item{Intercept_sd}{Standard deviation for the intercept prior (Normal). Default is NULL (flat).}

\item{b_sd}{Standard deviation for the coefficients prior (Normal). Default is NULL (flat).}

\item{sigma_rate}{Rate for the residual standard deviation prior (Exponential). Default is NULL (flat).}

\item{tau_rate}{Rate for the random effects standard deviation prior (Exponential). Default is NULL (flat).}

\item{...}{Optional hyperparameters}
}
\value{
A list with class "rtmb_prior"
}
\description{
Specify a uniform or manual prior
}
