% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_prior.R
\name{prior_uniform}
\alias{prior_uniform}
\title{Specify a flat prior}
\usage{
prior_uniform()
}
\value{
A list with class `"rtmb_prior"` and `type = "flat"`.
}
\description{
Deprecated alias of `prior_flat()`.
}
