% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{register_lpdf}
\alias{register_lpdf}
\title{カスタム尤度関数を lpdf 環境に登録する}
\usage{
register_lpdf(name, fun, force = FALSE)
}
\arguments{
\item{name}{登録する尤度関数の名前（文字列）}

\item{fun}{尤度を計算する関数}

\item{force}{すでに同名の関数が存在する場合に上書きするかどうか（デフォルトはFALSE）}
}
\description{
カスタム尤度関数を lpdf 環境に登録する
}
