% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{register_math}
\alias{register_math}
\title{Register a custom mathematical function to the math environment}
\usage{
register_math(name, fun, force = FALSE)
}
\arguments{
\item{name}{The name of the mathematical function to register (character string).}

\item{fun}{The function to register.}

\item{force}{Logical; whether to overwrite an existing function with the same name (default is FALSE).}
}
\description{
Register a custom mathematical function to the math environment
}
