% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{rtmb_code}
\alias{rtmb_code}
\title{Stan-like RTMB model definition block (Hybrid syntax supported)}
\usage{
rtmb_code(...)
}
\arguments{
\item{...}{Model definition blocks.}
}
\value{
A list of unevaluated code blocks.
}
\description{
Supports both R standard comma-separated syntax and Stan-like comma-less syntax (within {}).
}
