% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{rtmb_code}
\alias{rtmb_code}
\title{StanライクなRTMBモデル定義ブロック}
\usage{
rtmb_code(...)
}
\arguments{
\item{...}{各ブロックのコード (例: parameters = { ... })}
}
\value{
未評価のコードブロックのリスト
}
\description{
data, parameters, transform, model, generate の各ブロックをひとまとめにして記述します。
}
