% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{rtmb_corr}
\alias{rtmb_corr}
\title{相関行列 (多変量正規分布) 推定ラッパー}
\usage{
rtmb_corr(data, prior = list(lkj_eta = 1, mu_sd = 10, sigma_rate = 1))
}
\arguments{
\item{data}{観測データフレームまたは行列 (N x P)}

\item{prior}{事前分布のハイパーパラメータのリスト}
}
\description{
相関行列 (多変量正規分布) 推定ラッパー
}
