% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{rtmb_corr}
\alias{rtmb_corr}
\title{Wrapper for estimating correlation matrix (multivariate normal distribution)}
\usage{
rtmb_corr(
  data,
  prior = list(lkj_eta = 1, mu_sd = 10, sigma_rate = 1),
  init = NULL,
  null = NULL
)
}
\arguments{
\item{data}{Observation data frame or matrix (N x P).}

\item{prior}{List of hyperparameters for prior distributions. Default is \code{list(lkj_eta = 1.0, mu_sd = 10, sigma_rate = 1.0)}.}

\item{init}{List of initial values (optional).}

\item{null}{Target when creating a null model (e.g., \code{"corr"}). Optional.}
}
\value{
An instance of the \code{RTMB_Model} class.
}
\description{
Estimates a correlation matrix (along with means and standard deviations) assuming a multivariate normal distribution from observation data.
If there are 2 observed variables, it automatically switches to estimate the scalar correlation coefficient (`corr`) directly.
}
\examples{
inst/examples/ex_corr.R
}
