% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{rtmb_fa}
\alias{rtmb_fa}
\title{探索的因子分析 (Bayesian EFA) 安定・潜在変数モデル版ラッパー}
\usage{
rtmb_fa(
  data,
  nfactors = 1,
  rotate = NULL,
  score = FALSE,
  prior = list(mean_sd = 10, loadings_sd = 1, sd_rate = 10, ssp_ratio = 0.25),
  init = NULL
)
}
\arguments{
\item{data}{Observation data frame or matrix (N x J).}

\item{nfactors}{Number of factors (K).}

\item{rotate}{String specifying the rotation method (e.g., "varimax", "promax", "ssp"). If NULL, no rotation is applied. Specifying "ssp" performs regularized factor analysis.}

\item{score}{Logical; if TRUE, factor scores are calculated in the generate block (default is FALSE).}

\item{prior}{List of hyperparameters for prior distributions. `ssp_ratio` (or `loading_ratio`) represents the proportion of non-zero loadings per factor when "ssp" is specified.}

\item{init}{List of initial values. If not provided, initial values are automatically generated based on PCA or the psych package.}
}
\description{
探索的因子分析 (Bayesian EFA) 安定・潜在変数モデル版ラッパー
}
