% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{rtmb_fa}
\alias{rtmb_fa}
\title{探索的因子分析 (Bayesian EFA) 安定・潜在変数モデル版ラッパー}
\usage{
rtmb_fa(
  data,
  nfactors = 1,
  rotate = NULL,
  score = FALSE,
  prior = list(mean_sd = 10, loadings_sd = 1, sd_rate = 10),
  init = NULL
)
}
\arguments{
\item{data}{観測データフレームまたは行列 (N x P)}

\item{rotate}{回転方法の文字列 (例: "varimax", "promax")。NULLの場合は回転しない。}

\item{score}{論理値; TRUEの場合、generateブロックで因子得点を計算する (デフォルトはFALSE)}

\item{prior}{事前分布のハイパーパラメータのリスト}

\item{init}{初期値のリスト (指定しない場合はPCAに基づく初期値が自動生成されます)}

\item{n_factors}{因子の数 (K)}
}
\description{
探索的因子分析 (Bayesian EFA) 安定・潜在変数モデル版ラッパー
}
