% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{rtmb_fa}
\alias{rtmb_fa}
\title{RTMB-based Factor Analysis Wrapper}
\usage{
rtmb_fa(
  data,
  nfactors = 1,
  rotate = NULL,
  score = FALSE,
  prior = list(mean_sd = 10, loadings_sd = 1, sd_rate = 10, loading_ratio = 0.25),
  init = NULL
)
}
\arguments{
\item{data}{Observation data frame or matrix (N x J).}

\item{nfactors}{Number of factors (K).}

\item{rotate}{String specifying the rotation method (e.g., "varimax", "promax", "ssp"). If NULL, no rotation is applied. Specifying "ssp" performs regularized factor analysis.}

\item{score}{Logical; if TRUE, factor scores are calculated in the generate block (default is FALSE).}

\item{prior}{List of hyperparameters for prior distributions. `loading_ratio` represents the proportion of non-zero loadings per factor when "ssp" is specified.}

\item{init}{List of initial values. If not provided, initial values are automatically generated based on PCA or the psych package.}
}
\description{
Performs exploratory factor analysis (EFA) using RTMB. It supports standard
rotation methods (e.g., varimax, promax) as well as regularized factor analysis
using a Spike-and-Slab Prior (SSP) for estimating sparse loading matrices.
}
