% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{rtmb_fa}
\alias{rtmb_fa}
\title{探索的因子分析 (Bayesian EFA) 超高速版ラッパー}
\usage{
rtmb_fa(
  data,
  n_factors = 1,
  prior = list(mu_sd = 10, lambda_sd = 3, psi_rate = 1)
)
}
\arguments{
\item{data}{観測データフレームまたは行列 (N x P)}

\item{n_factors}{因子の数 (K)}

\item{prior}{事前分布のハイパーパラメータのリスト
探索的因子分析 (Bayesian EFA) 安定・潜在変数モデル版ラッパー}
}
\description{
探索的因子分析 (Bayesian EFA) 超高速版ラッパー
}
