% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{rtmb_glm}
\alias{rtmb_glm}
\title{RTMBベースのGLMラッパー関数 (変量効果なし)}
\usage{
rtmb_glm(
  formula,
  data,
  family = "gaussian",
  prior = list(beta_sd = 10, sigma_rate = 1, shape_rate = 1, phi_rate = 1)
)
}
\arguments{
\item{formula}{フォーミュラ (例: Y ~ X1 + X2)}

\item{data}{データフレーム}

\item{family}{分布族の文字列}

\item{prior}{事前分布のハイパーパラメータのリスト}
}
\description{
RTMBベースのGLMラッパー関数 (変量効果なし)
}
