% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{rtmb_glm}
\alias{rtmb_glm}
\title{RTMBベースのGLMラッパー関数 (変量効果なし)}
\usage{
rtmb_glm(
  formula,
  data,
  family = "gaussian",
  min_y = NULL,
  max_y = NULL,
  prior = list()
)
}
\arguments{
\item{formula}{フォーミュラ}

\item{data}{データフレーム}

\item{family}{分布族の文字列 ("gaussian", "binomial", "poisson" など)}

\item{min_y}{目的変数の理論上の最小値 (連続値モデルで推奨。省略時はデータから自動計算)}

\item{max_y}{目的変数の理論上の最大値 (連続値モデルで推奨。省略時はデータから自動計算)}

\item{prior}{事前分布のハイパーパラメータリスト}
}
\description{
RTMBベースのGLMラッパー関数 (変量効果なし)
}
