% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{rtmb_glmer}
\alias{rtmb_glmer}
\title{RTMBベースのGLMMラッパー関数}
\usage{
rtmb_glmer(
  formula,
  data,
  family = "gaussian",
  laplace = FALSE,
  min_y = NULL,
  max_y = NULL,
  prior = list(),
  init = NULL
)
}
\arguments{
\item{formula}{lme4スタイルのフォーミュラ (例: Y ~ X + (1 | GID))}

\item{data}{データフレーム}

\item{family}{分布族の文字列 ("gaussian", "binomial", "poisson" など)}

\item{laplace}{変量効果をラプラス近似で周辺化するかどうか}

\item{min_y}{目的変数の理論上の最小値 (連続値モデルで推奨。省略時はデータから自動計算)}

\item{max_y}{目的変数の理論上の最大値 (連続値モデルで推奨。省略時はデータから自動計算)}

\item{prior}{事前分布のハイパーパラメータリスト}

\item{init}{初期値のリスト (指定しない場合はglmに基づく初期値が自動生成されます)}
}
\description{
RTMBベースのGLMMラッパー関数
}
