% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{rtmb_glmer}
\alias{rtmb_glmer}
\title{RTMBベースのGLMMラッパー関数}
\usage{
rtmb_glmer(
  formula,
  data,
  family = "gaussian",
  laplace = FALSE,
  prior = list(beta_sd = 10, tau_rate = 1, sigma_rate = 1, lkj_eta = 1, shape_rate = 1,
    phi_rate = 1, cutpoint_sd = 2.5),
  init = NULL
)
}
\arguments{
\item{formula}{lme4スタイルのフォーミュラ (例: Y ~ X + (1 + X | GID))}

\item{data}{データフレーム}

\item{family}{分布族の文字列}

\item{laplace}{変量効果をラプラス近似で周辺化するかどうか (デフォルト: FALSE)}

\item{prior}{事前分布のハイパーパラメータのリスト}

\item{init}{初期値のリスト (指定しない場合はglmに基づく初期値が自動生成されます)}
}
\description{
RTMBベースのGLMMラッパー関数
}
