% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{rtmb_lm}
\alias{rtmb_lm}
\title{RTMBベースの線形回帰ラッパー関数}
\usage{
rtmb_lm(
  formula,
  data,
  min_y = NULL,
  max_y = NULL,
  prior = list(max_beta = 1, sd_scale = 2)
)
}
\arguments{
\item{formula}{フォーミュラ (例: Y ~ X1 + X2)}

\item{data}{データフレーム}

\item{min_y}{目的変数の理論上の最小値 (省略時はデータから自動計算)}

\item{max_y}{目的変数の理論上の最大値 (省略時はデータから自動計算)}

\item{prior}{事前分布のハイパーパラメータリスト (max_beta, sd_scale)}
}
\description{
RTMBベースの線形回帰ラッパー関数
}
