% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{rtmb_lm}
\alias{rtmb_lm}
\title{RTMB-based Linear Regression wrapper function}
\usage{
rtmb_lm(
  formula,
  data,
  y_range = NULL,
  penalty = c("none", "rhs", "ssp"),
  prior = list(prior_ratio = 1, max_beta = 1, logit_scale = 2.5, log_scale = 1,
    shape_rate = 1, phi_rate = 1, nu_rate = 0.1, cutpoint_sd = 2.5, expected_vars = 3,
    slab_scale = 2, slab_df = 4, ssp_ratio = 0.25),
  init = NULL
)
}
\arguments{
\item{formula}{Formula (e.g., Y ~ X1 + X2)}

\item{data}{Data frame}

\item{y_range}{Theoretical minimum and maximum values of the response variable as a vector c(min, max).}

\item{penalty}{Type of regularization for fixed effects: "none", "rhs" (Regularized Horseshoe), or "ssp" (Spike and Slab Prior). Default is "none".}

\item{prior}{List of hyperparameters for prior distributions}

\item{init}{List of initial values}
}
\description{
RTMB-based Linear Regression wrapper function
}
