% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{rtmb_model}
\alias{rtmb_model}
\title{Wrapper function to create an RTMB_Model instance}
\usage{
rtmb_model(
  data,
  parameters,
  model,
  transformed = NULL,
  generate = NULL,
  par_names = list(),
  init = NULL,
  view = NULL
)
}
\arguments{
\item{data}{A list of data used for modeling.}

\item{parameters}{A list (or alist) of parameters.}

\item{model}{A function or code generated by model_code().}

\item{transformed}{An optional function OR a list of expressions.}

\item{generate}{An optional function for generated quantities.}

\item{par_names}{A list of parameter names.}

\item{init}{A list or numeric vector of initial values.}

\item{view}{Character vector specifying the names of parameters to prioritize in the summary.}
}
\description{
Wrapper function to create an RTMB_Model instance
}
