% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{rtmb_model}
\alias{rtmb_model}
\title{Wrapper function to create an RTMB_Model instance}
\usage{
rtmb_model(
  data,
  code,
  par_names = list(),
  init = NULL,
  view = NULL,
  null_target = NULL
)
}
\arguments{
\item{data}{A named list containing observation data and constants (e.g., sample size) used in the model.}

\item{code}{A model definition block described by `rtmb_code(...)` (including data, parameters, model, transform, generate).}

\item{par_names}{A list of specific variable names corresponding to the dimensions of each parameter (optional).}

\item{init}{A list or numeric vector of initial values for parameters (optional). If not specified, initialized randomly.}

\item{view}{Character vector of parameter names to be displayed preferentially at the top when outputting results like `summary()` (optional).}

\item{null_target}{Character string. To simultaneously create a null model, specify the target parameter to fix and the prior distribution to disable as a formula string (e.g., \code{"delta ~ cauchy(0, r)"}). Default is \code{NULL}.}
}
\value{
An `RTMB_Model` class instance with compiled and pre-tested model.
}
\description{
Combines user-defined data and model code (likelihood and priors) to create an
`RTMB_Model` (R6 class) instance for Bayesian inference (MCMC, VB, MAP estimation).
}
