% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{rtmb_model}
\alias{rtmb_model}
\title{RTMB_Model インスタンスを生成するラッパー関数}
\usage{
rtmb_model(data, code, par_names = list(), init = NULL, view = NULL)
}
\arguments{
\item{data}{観測データを含む名前付きリスト}

\item{code}{`rtmb_code(...)` で記述されたモデル定義}

\item{par_names}{パラメータの変数名リスト}

\item{init}{初期値のリスト}

\item{view}{summaryで優先表示する変数名のベクトル}
}
\description{
RTMB_Model インスタンスを生成するラッパー関数
}
