% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{rtmb_model}
\alias{rtmb_model}
\title{RTMB_Modelインスタンスを生成するラッパー関数}
\usage{
rtmb_model(data, par_list, log_prob, generate = NULL)
}
\arguments{
\item{data}{モデリングに使用するデータのリスト}

\item{par_list}{Dim()で定義したパラメータのリスト}

\item{log_prob}{対数事後確率（または対数尤度＋事前分布）を計算する関数}

\item{generate}{生成量（Generated Quantities）を計算する関数（オプション）}
}
\value{
RTMB_Model クラスのインスタンス
}
\description{
RTMB_Modelインスタンスを生成するラッパー関数
}
