% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{sufficient_multi_normal_fa_lpdf}
\alias{sufficient_multi_normal_fa_lpdf}
\title{Sufficient statistics factor analysis multivariate normal log-probability density function}
\usage{
sufficient_multi_normal_fa_lpdf(S_mat, N, y_bar, mu, Lambda, psi)
}
\arguments{
\item{S_mat}{Deviation sum of squares matrix.}

\item{N}{Sample size.}

\item{y_bar}{Sample mean vector.}

\item{mu}{Mean parameter vector.}

\item{Lambda}{Factor loading matrix (P x K).}

\item{psi}{Vector of unique variances (P).}
}
\value{
The exact log-likelihood of the N raw observations.
}
\description{
Sufficient statistics factor analysis multivariate normal log-probability density function
}
