% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_post_estimation.R
\name{test_info}
\alias{test_info}
\title{Calculate Test Information Function}
\usage{
test_info(x, ...)
}
\arguments{
\item{x}{An object of class RTMB_Fit_Base}

\item{...}{Additional arguments.}
}
\description{
Calculate Test Information Function
}
\examples{
\dontrun{
  fit <- rtmb_irt(data = BigFive[, 1:5], model = "2pl")
  map_fit <- fit$optimize()
  ti <- test_info(map_fit)
  plot(ti)
}
}
