% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{transform_code}
\alias{transform_code}
\title{Transformed Code Wrapper for RTMB}
\usage{
transform_code(expr, env = parent.frame())

transform_code(expr, env = parent.frame())
}
\arguments{
\item{expr}{A block of code containing calculations for transformed parameters.}

\item{env}{Environment to assign to the generated function.}
}
\value{
A function taking (dat, par) that returns a named list.

A function taking (dat, par) that returns a named list.
}
\description{
Transformed Code Wrapper for RTMB

Transformed Code Wrapper for RTMB
}
